Improvements on message dispatching. Instead of doing fully async which does not guarantee sequentiality on xmpp messages we will use a single message dispatching thread initially backed by an unbouded queue.
The xmpp reading thread will continuously read and log xmpp messages and
will push these messages to the xmpp dispatcher thread. This dispatching
thread will invoke synchronously all the listeners and filters. 

Pontentially, a wrong application will be able to block the dispatching
queue by running a long filter/listener method, but that is left to the
framework (e.g. moho-remote) implementor. It is the framework
responsability to ensure listeners/filters are short-lived methods to
guarantee that messages are processed quickly.